requirements.txt
asgiref==3.3.1
Django==3.1.4
django-appconf==1.0.4
django-crispy-forms==1.10.0
django-redis==4.12.1
django-select2==7.5.0
Pillow==8.1.0
pkg-resources==0.0.0
pytz==2019.3
redis==3.5.3
sqlparse==0.3.1
